add docx2pdf option with to_pdf

+++ R/to_pdf.R
@@ -5,6 +5,11 @@
 #' can also be converted but some Word features are not supported
 #' such as sections.
 #' @param input,output file input and optional file output.
+#' @param use_docx2pdf if TRUE (and if 'Microsoft Word' executable
+#' can be found as well as 'docx2pdf'), docx2pdf will be used to
+#' convert 'Word' documents to PDF. This makes it possible to have a
+#' PDF identical to the 'Word' display whereas with 'LibreOffice', this
+#' is not always the case.
 #' @examples
 #' library(locatexec)
 #' if (exec_available("libreoffice")) {
@@ -18,7 +23,8 @@
 #' }
 #' @importFrom locatexec libreoffice_exec
 #' @importFrom locatexec is_windows
-to_pdf <- function(input, output = gsub("\\.[[:alnum:]]+$", ".pdf", input)) {
+to_pdf <- function(input, output = gsub("\\.[[:alnum:]]+$", ".pdf", input),
+                   use_docx2pdf = FALSE) {
   if (!file.exists(input)) {
     stop("input does not exist")
   }
@@ -32,30 +38,40 @@ to_pdf <- function(input, output = gsub("\\.[[:alnum:]]+$", ".pdf", input)) {
     to = default_root,
     overwrite = TRUE
   )
-  file_set_origin <- list.files(default_root, full.names = TRUE)
-  old_warn <- getOption("warn")
-  options(warn = -1)
-  info <- try(
-    system2(
-      libreoffice_exec(),
-      args = c("--headless",
-               if(!is_windows()) "\"-env:UserInstallation=file:///tmp/LibreOffice_Conversion_${USER}\"",
-               "--convert-to", "pdf:writer_pdf_Export",
-               "--outdir", shQuote(default_root, type = "cmd"),
-               shQuote(input, type = "cmd")),
-      stderr = TRUE, stdout = TRUE), silent = TRUE)
-  options(warn = old_warn)
-  out <- !1 %in% attr(info, "status")
-  if(!out) {
-    stop(paste0(info, collapse = "\n"))
+
+  if(grepl("\\.(doc|docx)$", input) && use_docx2pdf && exec_available("word") && docx2pdf_available()){
+    docx2pdf(input = input, output = output)
+  } else {
+    file_set_origin <- list.files(default_root, full.names = TRUE)
+    old_warn <- getOption("warn")
+    options(warn = -1)
+    info <- try(
+      system2(
+        libreoffice_exec(),
+        args = c("--headless",
+                 if(!is_windows()) "\"-env:UserInstallation=file:///tmp/LibreOffice_Conversion_${USER}\"",
+                 "--convert-to", "pdf:writer_pdf_Export",
+                 "--outdir", shQuote(default_root, type = "cmd"),
+                 shQuote(input, type = "cmd")),
+        stderr = TRUE, stdout = TRUE), silent = TRUE)
+    options(warn = old_warn)
+    out <- !1 %in% attr(info, "status")
+    if(!out) {
+      stop(paste0(info, collapse = "\n"))
+    }
+
+    file_set_new <- list.files(default_root, full.names = TRUE)
+    file_set_new <- setdiff(file_set_new, file_set_origin)
+    success <- file.copy(from = file_set_new, to = output, overwrite = TRUE)
+    unlink(file.path(default_root, basename(input)), force = TRUE)
+    unlink(file_set_new, force = TRUE)
+    if (any(!success)) {
+      stop("could not convert ", shQuote(input))
+    }
   }
 
-  file_set_new <- list.files(default_root, full.names = TRUE)
-  file_set_new <- setdiff(file_set_new, file_set_origin)
-  success <- file.copy(from = file_set_new, to = output, overwrite = TRUE)
-  unlink(file.path(default_root, basename(input)), force = TRUE)
-  unlink(file_set_new, force = TRUE)
-  if (any(!success)) stop("could not convert ", shQuote(input))
+
+
   output
 }
